requirements.txt
asgiref==3.2.5
backcall==0.1.0
beautifulsoup4==4.8.2
bootstrap4==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-appconf==1.0.3
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-heroku==0.3.1
django-imagekit==4.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
oauthlib==3.1.0
parso==0.6.2
pickleshare==0.7.5
pilkit==2.0
Pillow==7.0.0
prompt-toolkit==3.0.4
psycopg2==2.8.4
Pygments==2.6.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
whitenoise==5.0.1
